Changed uInt32 to size_t where necessary.
Finally decided to stop fighting C++, and just use size_t everywhere an
array is used.
Even though this is a 64-bit value, it seems that C++/STL expects it
when dealing with arrays, so I guess we will too.

+++ src/debugger/gui/CartDebugWidget.hxx
@@ -41,7 +41,7 @@ class CartDebugWidget : public Widget, public CommandSender
     virtual ~CartDebugWidget() = default;
 
   public:
-    int addBaseInformation(int bytes, const string& manufacturer,
+    int addBaseInformation(size_t bytes, const string& manufacturer,
         const string& desc, const uInt16 maxlines = 10);
 
     // Inform the ROM Widget that the underlying cart has somehow changed
